# swagger.json
{"swagger":"2.0","info":{"version":"v1","title":"Payment Gateway API","description":"Payment Gateway API"},"paths":{"/api/v{version}/payments":{"post":{"tags":["Payments"],"operationId":"ApiV{versionPaymentsPost","consumes":["application/json","text/json","application/json-patch+json"],"produces":[],"parameters":[{"name":"request","in":"body","required":false,"schema":{"$ref":"#/definitions/PaymentRequest"}},{"name":"version","in":"path","required":true,"type":"string"}],"responses":{"200":{"description":"Success"}}}},"/api/PaymentStatuses":{"get":{"tags":["PaymentStatuses"],"operationId":"ApiPaymentStatusesGet","consumes":[],"produces":["application/json"],"parameters":[{"name":"paymentId","in":"query","required":false,"type":"string"}],"responses":{"200":{"description":"Success","schema":{"$ref":"#/definitions/PaymentResponse"}}}}},"/api/v{version}/paymentstatus/{paymentId}":{"get":{"tags":["PaymentStatuses"],"operationId":"ApiV{versionPaymentstatusByPaymentIdGet","consumes":[],"produces":["application/json"],"parameters":[{"name":"paymentId","in":"path","required":true,"type":"string"},{"name":"version","in":"path","required":true,"type":"string"}],"responses":{"200":{"description":"Success","schema":{"$ref":"#/definitions/PaymentResponse"}}}}}},"definitions":{"PaymentRequest":{"type":"object","properties":{"id":{"type":"string"},"amount":{"format":"double","type":"number"},"currency":{"type":"string"},"merchantId":{"type":"string"},"createdDate":{"format":"date-time","type":"string"},"card":{"$ref":"#/definitions/Card"}}},"Card":{"type":"object","properties":{"number":{"type":"string"},"expiryMonth":{"format":"int32","type":"integer"},"expiryYear":{"format":"int32","type":"integer"},"cvv":{"type":"string"}}},"PaymentResponse":{"type":"object","properties":{"id":{"type":"string"},"paymentRequestId":{"type":"string"},"status":{"format":"int32","enum":[8000,8001],"type":"integer"},"statusCode":{"type":"string"},"card":{"$ref":"#/definitions/Card"}}}},"securityDefinitions":{}}